fixed some warnings

+++ Phobos/src/EditorLayer.cpp
@@ -954,7 +954,7 @@ namespace Ares
             if (ImGui::CollapsingHeader("Albedo", nullptr, ImGuiTreeNodeFlags_DefaultOpen))
             {
                 ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(10, 10));
-                ImGui::Image(m_AlbedoInput.TextureMap ? (void*)m_AlbedoInput.TextureMap->GetRendererID() : (void*)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
+                ImGui::Image(m_AlbedoInput.TextureMap ? (void*)(intptr_t)m_AlbedoInput.TextureMap->GetRendererID() : (void*)(intptr_t)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
                 ImGui::PopStyleVar();
                 if (ImGui::IsItemHovered())
                 {
@@ -964,7 +964,7 @@ namespace Ares
                         ImGui::PushTextWrapPos(ImGui::GetFontSize() * 35.0f);
                         ImGui::TextUnformatted(m_AlbedoInput.TextureMap->GetPath().c_str());
                         ImGui::PopTextWrapPos();
-                        ImGui::Image((void*)m_AlbedoInput.TextureMap->GetRendererID(), ImVec2(384, 384));
+                        ImGui::Image((void*)(intptr_t)m_AlbedoInput.TextureMap->GetRendererID(), ImVec2(384, 384));
                         ImGui::EndTooltip();
                     }
                     if (ImGui::IsItemClicked())
@@ -992,7 +992,7 @@ namespace Ares
             if (ImGui::CollapsingHeader("Normals", nullptr, ImGuiTreeNodeFlags_DefaultOpen))
             {
                 ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(10, 10));
-                ImGui::Image(m_NormalInput.TextureMap ? (void*)m_NormalInput.TextureMap->GetRendererID() : (void*)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
+                ImGui::Image(m_NormalInput.TextureMap ? (void*)(intptr_t)m_NormalInput.TextureMap->GetRendererID() : (void*)(intptr_t)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
                 ImGui::PopStyleVar();
                 if (ImGui::IsItemHovered())
                 {
@@ -1002,7 +1002,7 @@ namespace Ares
                         ImGui::PushTextWrapPos(ImGui::GetFontSize() * 35.0f);
                         ImGui::TextUnformatted(m_NormalInput.TextureMap->GetPath().c_str());
                         ImGui::PopTextWrapPos();
-                        ImGui::Image((void*)m_NormalInput.TextureMap->GetRendererID(), ImVec2(384, 384));
+                        ImGui::Image((void*)(intptr_t)m_NormalInput.TextureMap->GetRendererID(), ImVec2(384, 384));
                         ImGui::EndTooltip();
                     }
                     if (ImGui::IsItemClicked())
@@ -1021,7 +1021,7 @@ namespace Ares
             if (ImGui::CollapsingHeader("Metalness", nullptr, ImGuiTreeNodeFlags_DefaultOpen))
             {
                 ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(10, 10));
-                ImGui::Image(m_MetalnessInput.TextureMap ? (void*)m_MetalnessInput.TextureMap->GetRendererID() : (void*)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
+                ImGui::Image(m_MetalnessInput.TextureMap ? (void*)(intptr_t)m_MetalnessInput.TextureMap->GetRendererID() : (void*)(intptr_t)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
                 ImGui::PopStyleVar();
                 if (ImGui::IsItemHovered())
                 {
@@ -1031,7 +1031,7 @@ namespace Ares
                         ImGui::PushTextWrapPos(ImGui::GetFontSize() * 35.0f);
                         ImGui::TextUnformatted(m_MetalnessInput.TextureMap->GetPath().c_str());
                         ImGui::PopTextWrapPos();
-                        ImGui::Image((void*)m_MetalnessInput.TextureMap->GetRendererID(), ImVec2(384, 384));
+                        ImGui::Image((void*)(intptr_t)m_MetalnessInput.TextureMap->GetRendererID(), ImVec2(384, 384));
                         ImGui::EndTooltip();
                     }
                     if (ImGui::IsItemClicked())
@@ -1052,7 +1052,7 @@ namespace Ares
             if (ImGui::CollapsingHeader("Roughness", nullptr, ImGuiTreeNodeFlags_DefaultOpen))
             {
                 ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(10, 10));
-                ImGui::Image(m_RoughnessInput.TextureMap ? (void*)m_RoughnessInput.TextureMap->GetRendererID() : (void*)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
+                ImGui::Image(m_RoughnessInput.TextureMap ? (void*)(intptr_t)m_RoughnessInput.TextureMap->GetRendererID() : (void*)(intptr_t)m_CheckerboardTex->GetRendererID(), ImVec2(64, 64));
                 ImGui::PopStyleVar();
                 if (ImGui::IsItemHovered())
                 {
@@ -1062,7 +1062,7 @@ namespace Ares
                         ImGui::PushTextWrapPos(ImGui::GetFontSize() * 35.0f);
                         ImGui::TextUnformatted(m_RoughnessInput.TextureMap->GetPath().c_str());
                         ImGui::PopTextWrapPos();
-                        ImGui::Image((void*)m_RoughnessInput.TextureMap->GetRendererID(), ImVec2(384, 384));
+                        ImGui::Image((void*)(intptr_t)m_RoughnessInput.TextureMap->GetRendererID(), ImVec2(384, 384));
                         ImGui::EndTooltip();
                     }
                     if (ImGui::IsItemClicked())
@@ -1110,14 +1110,14 @@ namespace Ares
         ImGui::Begin("Toolbar");
         if (m_SceneState == SceneState::Edit)
         {
-            if (ImGui::ImageButton((ImTextureID)(m_PlayButtonTex->GetRendererID()), ImVec2(32, 32), ImVec2(0, 0), ImVec2(1, 1), -1, ImVec4(0, 0, 0, 0), ImVec4(0.9f, 0.9f, 0.9f, 1.0f)))
+            if (ImGui::ImageButton((ImTextureID)(intptr_t)(m_PlayButtonTex->GetRendererID()), ImVec2(32, 32), ImVec2(0, 0), ImVec2(1, 1), -1, ImVec4(0, 0, 0, 0), ImVec4(0.9f, 0.9f, 0.9f, 1.0f)))
             {
                 OnScenePlay();
             }
         }
         else if (m_SceneState == SceneState::Play)
         {
-            if (ImGui::ImageButton((ImTextureID)(m_PlayButtonTex->GetRendererID()), ImVec2(32, 32), ImVec2(0, 0), ImVec2(1, 1), -1, ImVec4(1.0f, 1.0f, 1.0f, 0.2f)))
+            if (ImGui::ImageButton((ImTextureID)(intptr_t)(m_PlayButtonTex->GetRendererID()), ImVec2(32, 32), ImVec2(0, 0), ImVec2(1, 1), -1, ImVec4(1.0f, 1.0f, 1.0f, 0.2f)))
             {
                 OnSceneStop();
             }
@@ -1175,7 +1175,7 @@ namespace Ares
 
         
         
-        ImGui::Image((void*)SceneRenderer::GetFinalColorBufferRendererID(), viewportSize, { 0, 1 }, { 1, 0 });
+        ImGui::Image((void*)(intptr_t)SceneRenderer::GetFinalColorBufferRendererID(), viewportSize, { 0, 1 }, { 1, 0 });
 
 
         /*m_GeoPass->GetSpecs().TargetFrameBuffer->Resize((uint32_t)viewportSize.x, (uint32_t)viewportSize.y);
